move type-specific stuff out of new graph visualizer
