+++ apps/hash-frontend/src/pages/shared/graph-visualizer.tsx
@@ -4,17 +4,25 @@ import { MultiDirectedGraph } from "graphology";
 import dynamic from "next/dynamic";
 import { memo, useState } from "react";
 
-import type { TypesGraphProps } from "./types-graph/graph-loader";
 import {
   FullScreenContextProvider,
   useFullScreen,
-} from "./types-graph/shared/full-screen";
+} from "./graph-visualizer/shared/full-screen";
+import type { GraphLoaderProps } from "./graph-visualizer/graph-loader";
+
+export type { GraphEdge, GraphNode } from "./graph-visualizer/graph-loader";
+
+export type GraphVisualizerProps = Omit<GraphLoaderProps, "highlightDepth">;
 
 const Graph = ({
   height,
-  onTypeClick,
-  types,
-}: Omit<TypesGraphProps, "highlightDepth"> & { height: string | number }) => {
+  edges,
+  nodes,
+  onEdgeClick,
+  onNodeClick,
+}: GraphVisualizerProps & {
+  height: number | string;
+}) => {
   /**
    * When a node is hovered or selected, we highlight its neighbors up to this depth.
    *
@@ -31,12 +39,14 @@ const Graph = ({
   );
 
   const TypesGraphLoader = dynamic(
-    import("./types-graph/graph-loader").then((module) => module.GraphLoader),
+    import("./graph-visualizer/graph-loader").then(
+      (module) => module.GraphLoader,
+    ),
     { ssr: false },
   );
 
   const FullScreenButton = dynamic(
-    import("./types-graph/full-screen-button").then(
+    import("./graph-visualizer/full-screen-button").then(
       (module) => module.FullScreenButton,
     ),
     { ssr: false },
@@ -52,26 +62,28 @@ const Graph = ({
       <FullScreenButton />
       <TypesGraphLoader
         highlightDepth={highlightDepth}
-        onTypeClick={onTypeClick}
-        types={types}
+        nodes={nodes}
+        edges={edges}
+        onEdgeClick={onEdgeClick}
+        onNodeClick={onNodeClick}
       />
     </SigmaContainer>
   );
 };
 
-export const TypesGraph = memo(
-  ({
-    height,
-    onTypeClick,
-    types,
-  }: Omit<TypesGraphProps, "highlightDepth"> & { height: string | number }) => {
+export const GraphVisualizer = memo(
+  (
+    props: GraphVisualizerProps & {
+      height: number | string;
+    },
+  ) => {
     /**
      * WebGL APIs aren't available in the server, so we need to dynamically load any module which uses Sigma/graphology.
      */
     if (typeof window !== "undefined") {
       return (
         <FullScreenContextProvider>
-          <Graph height={height} onTypeClick={onTypeClick} types={types} />
+          <Graph {...props} />
         </FullScreenContextProvider>
       );
     }
