+++ apps/hash-frontend/src/pages/shared/graph-visualizer/full-screen-button.tsx
