+++ apps/hash-frontend/src/pages/shared/graph-visualizer/shared/full-screen.tsx
